Tried to adjust code structure to iOS
The code structure should be somewhat similar to the iOS version now. If we really create a full Android version in the future, we should think over both codes, though. This is all somewhat too confusing and uses too many delegates and stuff.

Also, the web library is now notified of disconnected remotes, so applications should be fully working now.